Fix template logic
Parameter 'server_type' should be 'master' if the host is in the group
'thinlinc_masters', but not otherwise. This patch prevents the parameter
being set to 'master' even if the host is not in this group.

Commit 6410106 did not completely remove the requirement for having
specific group names in the inventory, and introduced a logic error.
With this commit the group requirement remains, but resolves the issue
where members of 'agent_servers' would be configured as masters.

Fixes #33.