Retire thinlinc_build variable
The ThinLinc build number can be useful to know for testing and
debugging purposes, but is almost never useful to specify explicitly.
This commit removes the need for this parameter by globbing to determine
which package(s) to install, rather than refering to them by their full
name (which includes the build number).

This also means we can eliminate a whole configuration file worth of
variables used for mapping architectures, versions and package formats
to their respective filenames. As part of this, we also remove the
pkgdir variable and infer the package locations from the version number
instead.

This commit also introduces an assertion to make sure we're not trying
to install ThinLinc >4.14.0 on a 32-bit system, since support for this
architecture has been removed.