Allow specifying bundle path
In fad51d0 we removed the thinlinc_server_bundle parameter. This means
that the bundle must be in a standard Ansible search path, using the
standard bundle filename.

This commit re-implements a variation of this parameter, but makes it
optional with sensible defaults. This makes it possible for people to
use non-standard locations and filenames, but still keeps the intention
of the original commit by only requiring the version to be specified in
one place.

Fixes #42.